[ThorpS3Client] Extract S3ClientUploader
